Update YouTube links for iOS
On iOS, the `UIWebView` has a property that governs whether media content
gets displayed inline or if the native full-screen controller is used.
If `allowsInlineMediaPlayback` is `TRUE` (iPad default), then inline is
allowed.  If `FALSE` (iPhone default), then attempts to display inline
are ignored.

Updated links so that once the LC code is updated, content will be
displayed inline on the iPhone.  Content will be displayed inline on an
iPad now due to the default property setting.