Fix Share button on browser views
When in a browser view (normal or landscape only), selecting the share
button would terminate the browser.  Added a check to ensure that the
app only performed the terminate when other buttons were selected.

Corrected a variable reference in the `closestack` handler of the
Siva-Siva-App stack script.

Added the ScriptTracker property to the binary stacks (so that the
plugin knows where to place the exported scripts).