Strict Compilation Mode Updates / DG Update
Updates to several scripts that will be touched to enable strict
compilation mode in the IDE.  Did find some questions where a variable
was not declared and looked like it may need to be a script local.

Updated the DG LayoutControl handler for the module list to size the
backdrop.  This will be needed when the DG width gets adjusted based
on the card width.

All changes to the binary stacks are reflected in the ScriptTracker
exports.