Responsive Design Updates
Add functions to return 'safeMargins'.  Initially this just supports the
iPhone X models to avoid the notch and widget at the bottom of the screen.

Update setCardCoords to allow the orientation to be forced.  When returning
from a landscape stack, the resizeStack message gets sent with the rect
still in landscape.  If an answer is put inside the resizeStack handler
then a second message is sent with the correct orientation.  Until another
method is determined, force the portal cards to portrait to avoid this issue.

Removed FSM rect designation from landscape only stacks (settings, wordpuzzles)

Changed portal stack and cards to responsive design.