Lexicon - responsive changes
Added `saveStackRequest` handler to clear the browser widget URL to avoid
errors when loading (Android at times will try to load the URL from the
dev environment before the preOpenStack can clear it).

On an iPhoneX, show the status bar.

Added function to get phone model to lib_UIHandlers - mainly to detect
iPhones that have a notch (just reported as iPhoneX for both sizes).

Grey areas at the top/bottom of the screen are part of the HTML5 code.