Add check boxes to set destroy stack/window on portal
For testing, added check boxes to set the destroy stack and destroy
window properties of the stack.

Initial results indicate that the existing setting works fine.  On an
Android Pixel, it changes the image that gets displayed when going back
(from browser, either a blank brown browser or the home page [existing
behavior])

Issue #235