Wordpuzzles nav issue workaround
When navigating to Word Puzzles stack, set the destroystack on Portal to
false for iPhones.  This is necessary to not have the shift of the main
nav screen after a noticeable delay.

When exiting the Word Puzzles stack, 3 options are possible:
1. Clear the puzzle and hide the buttons, background gets squished (lower button)
2. Show the green background used in the Portal, it is distorted (upper button)
3. Could also hide everything and use a background pattern (not implemented)

Fixed an issue where UI handler did not respect the forced orientation
in `setImageToFullCardLoc`.