Surprise responsive work / Nightly merge
Updated Siva-Siva-App and Siva-Siva-Portal to bring in the changes that
were made to the binary stacks in this branch.  They are updated and
ready to merge into nightly once the rest of the branch is ready.

Ensured that UI updated when cards opened
Positioned header
Positioned footer
Positioned buttons on surprise_verses card
Positioned go-next button
Positioned audioGlobalControl button
Positioned/sized background image

TODO:  Position of audio player on Android does not look correct on the
Pixel that is being used for testing.