WordPuzzle - alternate rotate fix
For some reason, iPhone would never show the next card before the screen
would lock prior to rotation.  There also was an issue when returning to
the puzzle stack where it ended up on the blank card.

Implemented an alternate fix.  Used a black graphic larger than the screen
size.  When the stack opens, be sure it is hidden.  When navigating back
to the portal, show the graphic and clear the puzzle.  Added a brief wait
on iPhone to be sure the screen was updated with the graphic prior to
doing the rotation.