Rotation Fix for Browser Views
Added `blankImage` grc to both browser view stacks.  Prior to closing
the stack, show the grc to prevent the previous content from being
scaled improperly in a rotation change.  Added a 10ms wait to the
landscape view (100ms already present in landscape) to ensure that the
grc is shown before the rotate takes place.

Ensure the grc is hidden in the preOpenStack handler of the behavior.

Removed the commented out code from the behavior stack.