Wordpuzzles bug fixes
On Android (Pixel), the order of the show in the reveal handler matters.
If the quote is not shown first, sections do not get drawn properly.
Issue did not effect IDE or iOS test.

If a clue ended up being an already selected word, it would glow and then
hang the script.  Added a check to only move an unlocked word.  Showed
the glow for only 500ms in that case.  Another option would be to skip
locked words as clues, but that would take more code.

If a clue was requested after the puzzle was complete, an error was
thrown.  Check to ensure the timer is still running when a clue is
requested.  Simply exit if the timer is stopped.

Updated the WordTile behavior to be strict compilation compliant.  Found
a variable name typo and corrected it.

Exported changes to external scripts (in this case they were all due to
object long ID changes) for ScriptTracker