Improve tree widget menu performance
Enclose redraw causing code inside a screen lock to prevent multiple
redraws.  This prevents the redraw that closes previous branch before
drawing the tree with the new branch expanded.

Move code that removes highlight to right after expand/collapse code.
This will minimize the time that the highlight is visible.