Rounding for center calculation
iPhone X, XS, and 11 Pro, 6, 6s, 7, 8 all have an odd screen width when
measured in points.  Currently, when centering an object that is the
width of the screen, it is placed one pixel off to the right.

This issue was initially noticed due to the footer in several modules that
have been made responsive and incorporated the new footer design.