Extend Footer into Chin Area
Most apps on iOS extend the footer background into the chin area.  To be consistent with other apps on the device, update the footer positioning code to do the same.  Since a change is needed everywhere that uses the footer, revised the code to send a message to the footer to update itself instead of doing the same action in each stack.

Surprise will still need to be updated in another commit.