Landscape to Portrait fix (iOS)
Add a one card "blank" stack to force rotation to happen properly on iOS.  When coming from wordpuzzles and the landscape browser, the call is redirected through the blank stack which forces the proper rotation behavior.  The way the call is currently written though will force all buttons on the browser stack to go to the portal.