Path To Siva Responsive Updates
Updated code to make remaining cards responsive.  Handlers located in the stack's script.
Replaced navigation with the footerNavigation group/behavior on the `choose your path` card
Added note to footerNavigation behavior - the group needs to be a background so that preOpenControl is sent prior to preOpenCard
Add `adjustLoc` handler to UI library