Fix footer crash
Footer must be a background for it to initialize correctly.  Added a `pass mouseUp` to the switch as a default action which should avoid the trapping that was happening.  Could possible rework the behavior to not need to be a background, but should do that separately from the loading update.