Implement lockTouches/unlockTouches
Moved the clearing of the highlight in the tree to give an indication when processing was still taking place.

In the portal, when checking to ensure a link is available (`checkURLAvailable`), lock the interface so that touches are discarded.  This prevents interaction with the UI if there is a delay in checking on the presence of the URL.  This happened in low bandwidth situations where a user could select other menu items while one was still being processed.

Corrected placement of showMandala image (size is smaller than the hide image)