Initial journal support, initial text sizing for story
Check to ensure the story text fills 75% of the field.  If not, then size up to 24pt larger.
Adjust journal code to return to correct page (set page to previous number so that goNext will land on correct page)
WIP: there is a long delay before "journalResume" is called in the IDE but the story does appear to be on the correct page.  Still looking for the reason.