Menu update, add pref to save story location
When viewing a server story, a pref is saved for the current scene.  When the user returns to the story, they are taken back to the last scene that was viewed.  At the end of the story, if the files are retained, then the count is reset back to the beginning.

Menu updated to only shade the stories yellow if they are not currently downloaded.