[[ lib_MobileControls ]] fix to update timer
The `currentTime` and `duration` properties of a player are not given in milliseconds as it was assumed by the code. They're given as an integer, and to get how many *seconds* they amount to, you need to divide them by the `playRate` property.

This makes the desktop code different than the mobile code which uses `mobileControlGet` to acquire values from the native player, all of which are given in milliseconds.

This code patches `updateTimer` and `convertTime` to make them work more reliable —without any run-away pending messages— and with the correct math for both environments.

Related to: #274 and #275