[[ listen ]] resuming paused audio is working now
The code was explicitly setting the value of `currenttime` when resuming a paused audio. The value being used was coming from `sTimestamp` which holds a value in milliseconds. This is not the correct value for desktop usage. The `currenttime` should be an integer between zero and the player's `duration` property, this value is not given in milliseconds.

This makes the mobile code use different values from the desktop code and makes debugging a lot harder.

This was fixed by not actually setting `currenttime` when resuming. There is no need for doing it. When you `start player`, it resumes from the last paused position automatically. All that timestamp handling between playing and pausing is not needed.

closes #276