[[ listen ]] Journal handling works
The journal handling has been fixed but this required changes to how the data is recorded in the journal. The timestamp can't simply be saved as `sTimestamp` and restored from it. It needs to be saved in the correct unit for that environment and restored directly by tweaking player properties.

Unfortunately, journal records already saved won't work as they have the wrong value for the timestamps (and the card, and the file id, and other properties)

closes: #278