[[ listen ]] url inserted in sDataA
In issue #284, aag notices that the URL value of `sDataA` is empty when resuming from the journal, even though it is explicitly filled in in `journalResume`. Some other handler somewhere is clobbering `sDataA`. This patch makes sure that when any track starts playing, that the URL value of `sDataA` is filled back in.