[[ listen ]] fix for updating local Jnanam with data from server-side playlists
Created a new command in `Jnanam` library called `jnanam_SaveNewItem`. This command only inserts a new item if the `file_id` of the item it is attempting to insert is not on the database already.

Used this command in both `playlist_FetchData` to loop through all the items in `sCurrentPlaylistA` inserting them as needed and also on `downloadAffirmed` to make sure that the item for the audio being downloaded is on the local database by the time the download finishes (this is probably unneeded but I did it before doing the other one).

This fixes #294