Fixed scroller issue on Listen
The "CreateScroller" was being called only for the "listen-browse" card.  Moved the command to where it will be called for the "listen-my-audio" card as well.  Also made the call work for both cards.

"CreateScroller" was only being called when field was too big to display the whole list.  Since we are capturing the tap using the scroller code, needed to always create a scroller overlay.

Did the ScriptTracker exports.

Removed excess code (commented out)

Corrected field reference in fld "searchString"