[colorIt] rotate bug fix
When leaving colorIt from landscape, on iOS the screen would get stuck drawn for portrait but displayed on the device in landscape.  Removed the orientation lock which wasn't needed and added a call to force portrait and then a wait to ensure it had time to take effect before releasing control to the user.

Removed a debugging put that is no longer needed.