[stories] fix iOS image rotation
If a mobile scroller for a full screen image is set up too soon after rotating, it causes the app to lock up and become unresponsive on iOS.  Solution is to send the scroller create in time to happen 10ms after doing all of the rotate adjustments.  This restores the same functionality to iOS that Android had (could rotate even if the image was in a zoomed state).